refactor: rename *Rule -> *WalletRule
