+++ packages/account-tree-controller/src/rules/index.ts
@@ -1,4 +1,4 @@
-export * from './Rule';
-export * from './EntropySourceRule';
-export * from './SnapRule';
-export * from './KeyringTypeRule';
+export * from './WalletRule';
+export * from './EntropySourceWalletRule';
+export * from './SnapWalletRule';
+export * from './KeyringWalletRule';
